Initializer should not have `func` keyword
I actually thought of declaring `isInitializer` like `let isInitializer = ["init", "init?"].contains(rawName)`
But this is kind of different issue. Failable initializer looses question mark when generated as a class method name.
So this code `init?(a: Int) { return nil }` will result with token.name like this `init(a:)`
Not yet sure if this is an issue though.